+++ sdk/storage/Azure.Storage.DataMovement/CHANGELOG.md
@@ -6,6 +6,7 @@
 - `TransferManager` new API `PauseAllRunningTransfersAsync`.
 - Added support for `TransferManager.GetTransfers`, to retrieve the list of transfers in the `TransferManager`.
 - Added support for tracking progress of transfers. See `TransferOptions.ProgressHandler` and `TransferOptions.ProgressHandlerOptions`.
+- Added support for `Transfermanager.ResumeTransferAsync` to resume a transfer.
 
 ### Breaking Changes
 - [BREAKING CHANGE] Altered API signatures on `TransferManager` and `DataTransfer` for pausing.